# sdks/dotnet/api/openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: hello@zahidaramai.com
    name: KLCUBE NETWORK Enterprise
  description: |
    API middleware gateway for Malaysia's MyInvois e-invoicing system (LHDN).

    **Client:** Hashmato | **Version:** 1.1.1

    ## Features
    - 4 Submission Endpoints (Consolidate, JustSave, Buyer, Personal)
    - JWT Authentication with Access & Refresh Tokens
    - User/Role/Company Management
    - Document Operations (List, Status, PDF, Cancel)
    - Digital Signing v1.1 (XAdES signatures)
    - Background Status Polling

    ## Authentication
    All endpoints except `/api/v1/auth/login` require JWT authentication.
    Include the access token in the Authorization header:
    ```
    Authorization: Bearer <access_token>
    ```
  title: HashLHDN MyInvois Middleware API
  version: 1.1.1
servers:
  - description: Local Development
    url: http://localhost:3000
  - description: Production
    url: https://api.hashlhdn.com
...
paths:
...
  /api/v1/documents/{uuid}/cancel:
    post:
      description: Cancel a submitted document (within 72 hours)
      operationId: cancelDocument
      parameters:
        - explode: false
          in: path
          name: uuid
          required: true
          schema:
            format: uuid
            type: string
          style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/cancelDocument_request"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/cancelDocument_200_response"
          description: Document cancelled
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Invalid request
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Resource not found
        "429":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Rate limit exceeded
          headers:
            Retry-After:
              description: Seconds to wait before retrying
              explode: false
              schema:
                type: integer
              style: simple
      security:
        - bearerAuth: []
      summary: Cancel document
      tags:
        - Documents
...
components:
...
  schemas:
    ErrorResponse:
      example:
        error:
          code: code
          details:
            - "{}"
            - "{}"
          message: message
      properties:
        error:
          $ref: "#/components/schemas/ErrorResponse_error"
      required:
        - error
      type: object
...
    cancelDocument_request:
      properties:
        reason:
          maxLength: 500
          minLength: 1
          type: string
      required:
        - reason
      type: object
    cancelDocument_200_response:
      example:
        message: message
        uuid: uuid
        status: CANCELLED
      properties:
        uuid:
          type: string
        status:
          example: CANCELLED
          type: string
        message:
          type: string
      type: object
...
    ErrorResponse_error:
      example:
        code: code
        details:
          - "{}"
          - "{}"
        message: message
      properties:
        code:
          type: string
        message:
          type: string
        details:
          items:
            type: object
          type: array
      required:
        - code
        - message
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http